- написал тесты для репозиториев

+++ src/main/java/edu/skypro/coursework/examquestionsapp/services/impl/JavaQuestionService.java
@@ -1,6 +1,7 @@
 package edu.skypro.coursework.examquestionsapp.services.impl;
 
 import edu.skypro.coursework.examquestionsapp.exceptions.InvalidInputException;
+import edu.skypro.coursework.examquestionsapp.exceptions.QuestionNotFoundException;
 import edu.skypro.coursework.examquestionsapp.model.Question;
 import edu.skypro.coursework.examquestionsapp.repositories.QuestionRepository;
 import edu.skypro.coursework.examquestionsapp.services.QuestionService;
@@ -32,6 +33,9 @@ public class JavaQuestionService implements QuestionService {
 
     @Override
     public Question remove(Question question) {
+        if (!repository.getAll().contains(question)) {
+            throw new QuestionNotFoundException();
+        }
         return repository.remove(question);
     }
 
@@ -46,6 +50,6 @@ public class JavaQuestionService implements QuestionService {
         return repository.getAll().stream()
                 .skip(random.nextInt(repository.getAll().size()))
                 .findFirst()
-                .orElseThrow(InvalidInputException::new);
+                .orElseThrow(QuestionNotFoundException::new);
     }
 }
